pyproject.toml
[tool.poetry]
name = "pyraft"
version = "0.1.0"
description = ""
authors = ["wang_chao03 <wang_chao03@inspur.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
pycryptodome = "^3.17"
msgpack = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
